Update ApprParabola2.h
If meanSquareError is null in the call to Fit, the local vecU is not copied to the output u. Thanks to Julien Maille for reporting and fixing the bug. The same bug occurs in FitRobust and needed to be fixed. The latest code has both fixes. The previous code uses a linear system solver specialized for 3x3 matrix systems. The fixes use a linear system solver with general purpose Gaussian elimination. This will result in lightly slower performance, but could be costly if you call the fitter a large number of times. The ApprParaboloid3 class has similar logic and uses the general purpose Gaussian elimination solely because GTE does not contain a specialization for 6x6 matrix systems. (I have added to my TODO list to use only specialized matrix system solvers for the
fitting.)