Update to OpenGL support, fix to GenerateProject.
After removing support for MSVS 2019, some files in the GenerateProject tools were deleted. However, I had failed to remove the project references to those files, and the project failed to compile.

For the OpenGL trace system, I added wrappers to hide the Microsoft OpenGL 1.0 and 1.1 DLL functions. This is a hack for now which uses the gte namespace. Fighting the gl* names for the DLL functions when the remaining extensions have wrappers around them with gl* names was painful. A later modification of GTE OpenGL will rename all the wrappers Gl* to avoid the hack.