Update GL46.cpp
The function OpenGLReportListener was incorrect regarding throwing exceptions. The throw occurred inside the 'default' case of the switch statement. It had to be outside the switch statement and inside the code block for when 'code' is not zero. After making this change, the Intel Graphics driver on my laptop threw an exception in the GLSL reflection code when checking for a GL error in glProgramInterfaceiv for programInterface is GL_TRANFORM_BUFFER. I have seen this behavior over the years. I thought that the OpenGL specification required this to be handled. I added a block of code to trap this special case and not call \textsf{ReportGLError}. I also commented out the preprocessor definition GTE_GL46_THROW_ON_REPORT_LISTENER_WARNING}
in case other such GL errors are unexpectedly generated.