Update TriangulateEC.h
The function ComputeNearestOuterPolygonIntersection had an assertion that v0min and v1min were nonnegative. However, they are of type std::size_t, so the assertion is never triggered. This problem was introduced in version 6.0.2023.12.16 when I converted from type
std::int32_t to std::size_t. Thanks to Aaron Barany for diagnosing and reporting the problem. I am working on modifications to trap more problems when the input outer and inner polygons are invalid.