Fixed assertion in ExtremalQuery sample application when BSP queries are used.
I had claimed previously that the ExtremalQuery3BSP code was broken. It turns out that the code works as before. The problem was that I had modified the ConvexHull3 code whereby the old code function GetUnorderedHull returned an array of triangles but the new code function GetHull returned an array of indices. The application code was not modified accordingly, so when the application was set to use the BSP query, it would trigger an assert. (I need the BSP query to fix a performance problem I introduced in a bug fix to MinimumVolumeBox3.) The query header was modified to avoid the Visual Studio code analysis warnings that the SphericalArc objects were possibly uninitialized, even though the default constructor initializes all member functions. The application now uses the BSP query by default.