Fixed hill-climbing performance and redesign of classes.
The GetExtreme function is a hill-climbing algorithm that visits the vertices of a convex hull. A user reported failure of the minimum volume box code, which I tracked down to failure of GetExtreme. I posted a fix on July 24, 2024. Another dataset showed this did not solve the original problem. I added another fix on August 31, 2024. This also did not solve the original problem. The final fix was to skip hill climbing and search all vertices, posted on September 5, 2024. Over 1 year later, another user of the code reported that the performance degraded significantly. The original datasets that showed failure had a very small number of points. The other user had datasets with thousands of points.

I spent quite a bit of time doing a deep dive of the code. It turns out the actual problem was that a convex hull can have triangular faces where some of the faces are coplanar (or nearly coplanar). Failure can occur when a vertex interior to a union of coplanar faces is visited. The algorithm terminated prematurely. My BSP-based extreme point code failed for the same reason. The minimum volume box code maintains a vertex-adjacency data structure built directly from the convex hull. The solution to the original problem was to remove edges from this data structure that are shared by two coplanar triangles. After such removal, there can be some vertices remaining that have only 2 neighbors. These also had to be removed. Effectively, the data structure represents a wireframe of the convex hull where some of the faces
are convex polygons with 4 or more vertices.

I redesigned the class interface slightly, separating out the version that uses floating-point computations and the version that uses rational arithmetic. Both versions must use rational arithmetic to build the convex hull for an exact representation. Without this, the hill climbing can fail. Fortunately, the rational convex hull is not expensive to compute given that the dot products in GetExtreme are the main bottleneck in the computations.