Posting GTE 8.1.
GaussianElimination.h: The private Set function had an if-else statement that was designed to provide fast copying for floating-point Real and object copying for rational Real. The if-else statement  provides flow control at run time. However, the separation must occur at compile time. The Ubuntu g++ 13.3.0 compiler generated an error for the separation. I modified the code to use std::enable_if and  provided a Set for each case, the other case ignored by SFINAE (Substitution Failure Is Not An Error).

MinimumVolumeBox3*.h: The g++ compiler warned about the use of a pragma that is supported by the Microsoft Visual Studio compiler. With treat-warnings-as-errors, the files do not compile. I removed the pragma statements. The g++ compiler also complained about the Edge.T and Triangle.T members shadowing the class template parameter T. I renamed the members of Edge and Triangle to use lower-case v, e, and t.

MinimumVolumeBox3DWindow3.cpp: The modifications to the minimum volume box code required an update to the sample application. The new template parameter list is different from the old one.